chore: refactor metrics collector code
